requirements.txt
selenium
allure-behave==2.10.0
allure-python-commons==2.10.0
behave==1.2.6
webdriver-manager==3.8.3

